pyproject.toml
[project]
name = "vectors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "asyncio>=3.4.3",
    "chonkie>=1.1.0",
    "docling>=2.39.0",
    "duckdb>=1.3.2",
    "faster-whisper>=1.1.1",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-docling>=1.0.0",
    "langchain-huggingface>=0.3.0",
    "langchain-milvus>=0.2.0",
    "langchain-ollama>=0.3.3",
    "langchain-postgres>=0.0.15",
    "numpy>=2.3.1",
    "oracledb>=3.2.0",
    "pandas>=2.3.0",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary,pool]>=3.2.9",
    "replicate>=1.0.7",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
    "transformers>=4.53.0",
    "zlib-ng>=0.5.1",
]
